fix result_wrapper.hpp(309): warning C4305: '=': truncation from 'int' to 'bool' [C:\projects\metaSMT\build\tests\direct_Z3.vcxproj]

+++ src/metaSMT/result_wrapper.hpp
@@ -306,7 +306,7 @@ namespace metaSMT {
           Integer ret = 0;
           bool isSigned = boost::is_signed<Integer>::value && val[val.size()-1];
           std::cout << "isSigned: " << isSigned << std::endl;
-          if( isSigned ) ret = -1 ;
+          if( isSigned ) ret = static_cast<Integer>(-1) ;
           for (unsigned i = 0; i < val.size(); ++i) {
             ret ^= Integer( val[i]^(isSigned ? static_cast<Integer>(1) : static_cast<Integer>(0))) << i;
           }
